add display sleep command and additional sy6970 functionality

+++ firmware/src/charge/charge_driver.h
@@ -20,14 +20,14 @@ typedef struct {
 
 typedef struct {
   uint16_t voltage_mv; // Voltage in millivolts
-  float percentage;
+  uint8_t percentage;  // Percentage of battery charge
 } VoltagePoint;
 
 // Generic interface for charge driver
 char *charge_state_to_string(RemoteChargeState state);
 esp_err_t charge_driver_init();
 RemotePowerState get_power_state();
-float battery_mv_to_percent(uint16_t voltage_mv);
+uint8_t battery_mv_to_percent(uint16_t voltage_mv);
 void disable_watchdog();
 void enable_watchdog();
